refactor: move template generator to it's own library
The code behind generating zip templates has been refactored into it's
own npm package to allow for us to test it and reuse it for a standalone
CLI tool. Sadly, because NodeJS doesn't have a usable fetch (yet) and
the browser has to deal with CORs we don't make any opinion on how
`gradle-wrapper.jar` should be loaded, instead opting to leave that to
the consumer of the library to handle themselves.